+++ hypnotoad/gui/hypnotoad_mainWindow.py
@@ -115,6 +115,10 @@ class Ui_Hypnotoad(object):
         self.action_Preferences.setObjectName(u"action_Preferences")
         icon9 = QIcon(QIcon.fromTheme(u"document-properties"))
         self.action_Preferences.setIcon(icon9)
+        self.action_Regrid = QAction(Hypnotoad)
+        self.action_Regrid.setObjectName(u"action_Regrid")
+        self.action_Regrid.setEnabled(False)
+        self.action_Regrid.setIcon(icon6)
         self.centralwidget = QWidget(Hypnotoad)
         self.centralwidget.setObjectName(u"centralwidget")
         self.horizontalLayout_2 = QHBoxLayout(self.centralwidget)
@@ -139,50 +143,72 @@ class Ui_Hypnotoad(object):
 
         self.verticalLayout_2.addWidget(self.options_form)
 
-        self.horizontalLayout_3 = QHBoxLayout()
-        self.horizontalLayout_3.setObjectName(u"horizontalLayout_3")
-        self.options_file_label = QLabel(self.centralwidget)
-        self.options_file_label.setObjectName(u"options_file_label")
-
-        self.horizontalLayout_3.addWidget(self.options_file_label)
-
-        self.options_file_line_edit = QLineEdit(self.centralwidget)
-        self.options_file_line_edit.setObjectName(u"options_file_line_edit")
-
-        self.horizontalLayout_3.addWidget(self.options_file_line_edit)
-
+        self.gridLayout_2 = QGridLayout()
+        self.gridLayout_2.setObjectName(u"gridLayout_2")
         self.options_file_browse_button = QPushButton(self.centralwidget)
         self.options_file_browse_button.setObjectName(u"options_file_browse_button")
 
-        self.horizontalLayout_3.addWidget(self.options_file_browse_button)
-
-        self.geqdsk_file_label = QLabel(self.centralwidget)
-        self.geqdsk_file_label.setObjectName(u"geqdsk_file_label")
-
-        self.horizontalLayout_3.addWidget(self.geqdsk_file_label)
-
-        self.geqdsk_file_line_edit = QLineEdit(self.centralwidget)
-        self.geqdsk_file_line_edit.setObjectName(u"geqdsk_file_line_edit")
-
-        self.horizontalLayout_3.addWidget(self.geqdsk_file_line_edit)
+        self.gridLayout_2.addWidget(self.options_file_browse_button, 0, 2, 1, 1)
 
         self.geqdsk_file_browse_button = QPushButton(self.centralwidget)
         self.geqdsk_file_browse_button.setObjectName(u"geqdsk_file_browse_button")
 
-        self.horizontalLayout_3.addWidget(self.geqdsk_file_browse_button)
+        self.gridLayout_2.addWidget(self.geqdsk_file_browse_button, 0, 6, 1, 1)
+
+        self.options_file_line_edit = QLineEdit(self.centralwidget)
+        self.options_file_line_edit.setObjectName(u"options_file_line_edit")
+
+        self.gridLayout_2.addWidget(self.options_file_line_edit, 0, 1, 1, 1)
 
         self.run_button = QPushButton(self.centralwidget)
         self.run_button.setObjectName(u"run_button")
 
-        self.horizontalLayout_3.addWidget(self.run_button)
+        self.gridLayout_2.addWidget(self.run_button, 0, 7, 1, 1)
 
         self.write_grid_button = QPushButton(self.centralwidget)
         self.write_grid_button.setObjectName(u"write_grid_button")
 
-        self.horizontalLayout_3.addWidget(self.write_grid_button)
+        self.gridLayout_2.addWidget(self.write_grid_button, 0, 9, 1, 1)
+
+        self.geqdsk_file_line_edit = QLineEdit(self.centralwidget)
+        self.geqdsk_file_line_edit.setObjectName(u"geqdsk_file_line_edit")
+
+        self.gridLayout_2.addWidget(self.geqdsk_file_line_edit, 0, 4, 1, 1)
+
+        self.geqdsk_file_label = QLabel(self.centralwidget)
+        self.geqdsk_file_label.setObjectName(u"geqdsk_file_label")
+
+        self.gridLayout_2.addWidget(self.geqdsk_file_label, 0, 3, 1, 1)
+
+        self.options_file_label = QLabel(self.centralwidget)
+        self.options_file_label.setObjectName(u"options_file_label")
+
+        self.gridLayout_2.addWidget(self.options_file_label, 0, 0, 1, 1)
+
+        self.regrid_button = QPushButton(self.centralwidget)
+        self.regrid_button.setObjectName(u"regrid_button")
+        self.regrid_button.setEnabled(False)
+
+        self.gridLayout_2.addWidget(self.regrid_button, 1, 7, 1, 1)
+
+        self.nonorthogonal_box = QCheckBox(self.centralwidget)
+        self.nonorthogonal_box.setObjectName(u"nonorthogonal_box")
+        self.nonorthogonal_box.setEnabled(True)
+
+        self.gridLayout_2.addWidget(self.nonorthogonal_box, 1, 6, 1, 1)
 
 
-        self.verticalLayout_2.addLayout(self.horizontalLayout_3)
+        self.verticalLayout_2.addLayout(self.gridLayout_2)
+
+        self.horizontalLayout_4 = QHBoxLayout()
+        self.horizontalLayout_4.setObjectName(u"horizontalLayout_4")
+
+        self.verticalLayout_2.addLayout(self.horizontalLayout_4)
+
+        self.gridLayout = QGridLayout()
+        self.gridLayout.setObjectName(u"gridLayout")
+
+        self.verticalLayout_2.addLayout(self.gridLayout)
 
 
         self.horizontalLayout.addLayout(self.verticalLayout_2)
@@ -212,6 +238,12 @@ class Ui_Hypnotoad(object):
         self.toolBar = QToolBar(Hypnotoad)
         self.toolBar.setObjectName(u"toolBar")
         Hypnotoad.addToolBar(Qt.TopToolBarArea, self.toolBar)
+        self.toolBar_2 = QToolBar(Hypnotoad)
+        self.toolBar_2.setObjectName(u"toolBar_2")
+        Hypnotoad.addToolBar(Qt.TopToolBarArea, self.toolBar_2)
+        self.toolBar_3 = QToolBar(Hypnotoad)
+        self.toolBar_3.setObjectName(u"toolBar_3")
+        Hypnotoad.addToolBar(Qt.TopToolBarArea, self.toolBar_3)
 
         self.menubar.addAction(self.menu_File.menuAction())
         self.menubar.addAction(self.menu_Mesh.menuAction())
@@ -226,6 +258,7 @@ class Ui_Hypnotoad(object):
         self.menu_File.addAction(self.action_Quit)
         self.menu_Help.addAction(self.action_About)
         self.menu_Mesh.addAction(self.action_Run)
+        self.menu_Mesh.addAction(self.action_Regrid)
         self.menu_Mesh.addAction(self.action_Write_grid)
         self.toolBar.addAction(self.action_New)
         self.toolBar.addAction(self.action_Open)
@@ -275,18 +308,26 @@ class Ui_Hypnotoad(object):
 #endif // QT_CONFIG(shortcut)
         self.action_Revert.setText(QCoreApplication.translate("Hypnotoad", u"&Revert", None))
         self.action_Preferences.setText(QCoreApplication.translate("Hypnotoad", u"&Preferences...", None))
+        self.action_Regrid.setText(QCoreApplication.translate("Hypnotoad", u"Re&grid", None))
+        self.action_Regrid.setToolTip(QCoreApplication.translate("Hypnotoad", u"Regrid non-orthogonal mesh", None))
         ___qtablewidgetitem = self.options_form.horizontalHeaderItem(0)
         ___qtablewidgetitem.setText(QCoreApplication.translate("Hypnotoad", u"Name", None));
         ___qtablewidgetitem1 = self.options_form.horizontalHeaderItem(1)
         ___qtablewidgetitem1.setText(QCoreApplication.translate("Hypnotoad", u"Value", None));
-        self.options_file_label.setText(QCoreApplication.translate("Hypnotoad", u"Options file", None))
         self.options_file_browse_button.setText(QCoreApplication.translate("Hypnotoad", u"Browse", None))
-        self.geqdsk_file_label.setText(QCoreApplication.translate("Hypnotoad", u"geqdsk file", None))
         self.geqdsk_file_browse_button.setText(QCoreApplication.translate("Hypnotoad", u"Browse", None))
         self.run_button.setText(QCoreApplication.translate("Hypnotoad", u"Run", None))
         self.write_grid_button.setText(QCoreApplication.translate("Hypnotoad", u"Write Grid", None))
+        self.geqdsk_file_label.setText(QCoreApplication.translate("Hypnotoad", u"geqdsk file", None))
+        self.options_file_label.setText(QCoreApplication.translate("Hypnotoad", u"Options file", None))
+        self.regrid_button.setToolTip(QCoreApplication.translate("Hypnotoad", u"Recalculate spacing of non-orthogonal grids (check 'Non-Orthogonal' box to activate)", None))
+        self.regrid_button.setText(QCoreApplication.translate("Hypnotoad", u"Regrid", None))
+        self.nonorthogonal_box.setToolTip(QCoreApplication.translate("Hypnotoad", u"Check to generate non-orthogonal grids", None))
+        self.nonorthogonal_box.setText(QCoreApplication.translate("Hypnotoad", u"Non-Orthogonal", None))
         self.menu_File.setTitle(QCoreApplication.translate("Hypnotoad", u"&File", None))
         self.menu_Help.setTitle(QCoreApplication.translate("Hypnotoad", u"&Help", None))
         self.menu_Mesh.setTitle(QCoreApplication.translate("Hypnotoad", u"&Mesh", None))
         self.toolBar.setWindowTitle(QCoreApplication.translate("Hypnotoad", u"toolBar", None))
+        self.toolBar_2.setWindowTitle(QCoreApplication.translate("Hypnotoad", u"toolBar_2", None))
+        self.toolBar_3.setWindowTitle(QCoreApplication.translate("Hypnotoad", u"toolBar_3", None))
     # retranslateUi